// frontend/src/types/Company.ts
export type TCompany = {
    id: string;
    name: string;
    description: string;
    logo: string;
    city: string;
    country: string;
    streetAddress: string;
    fullAddress: string;
    zipCode: number;
    orgNr: number;
    bankNr: number;
    phoneNumber: number;
    latitude: number;
    longitude: number;
    contactEmail: string;
    invoiceEmail: string;
    created: Date;
}